Move and rename assignment to course_assignments_metrics
This better reflects what the endpoint returns.
