+++ lms/routes.py
@@ -269,11 +269,12 @@ def includeme(config):  # noqa: PLR0915
         "api.dashboard.assignment", "/api/dashboard/assignments/{assignment_id}"
     )
     config.add_route("api.dashboard.course", "/api/dashboard/courses/{course_id}")
-    config.add_route(
-        "api.dashboard.course.assignments.stats",
-        "/api/dashboard/courses/{course_id}/assignments/stats",
-    )
+
     config.add_route("api.dashboard.assignments", "/api/dashboard/assignments")
+    config.add_route(
+        "api.dashboard.course.assignments.metrics",
+        "/api/dashboard/courses/{course_id}/assignments/metrics",
+    )
 
     config.add_route("api.dashboard.students", "/api/dashboard/students")
     config.add_route(
