+++ lms/views/dashboard/api/assignment.py
@@ -2,11 +2,12 @@ from marshmallow import fields, validate
 from pyramid.view import view_config
 
 from lms.js_config_types import (
+    AnnotationMetrics,
     APIAssignment,
     APIAssignments,
     APICourse,
 )
-from lms.models import Assignment
+from lms.models import Assignment, RoleScope, RoleType
 from lms.security import Permissions
 from lms.services.h_api import HAPI
 from lms.views.dashboard.pagination import PaginationParametersMixin, get_page
@@ -25,6 +26,7 @@ class AssignmentViews:
         self.h_api = request.find_service(HAPI)
         self.assignment_service = request.find_service(name="assignment")
         self.dashboard_service = request.find_service(name="dashboard")
+        self.course_service = request.find_service(name="course")
 
     @view_config(
         route_name="api.dashboard.assignments",
@@ -62,3 +64,54 @@ class AssignmentViews:
             title=assignment.title,
             course=APICourse(id=assignment.course.id, title=assignment.course.lms_name),
         )
+
+    @view_config(
+        route_name="api.dashboard.course.assignments.metrics",
+        request_method="GET",
+        renderer="json",
+        permission=Permissions.DASHBOARD_VIEW,
+    )
+    def course_assignments_metrics(self) -> APIAssignments:
+        course = self.dashboard_service.get_request_course(self.request)
+        course_students = self.course_service.get_members(
+            course, role_scope=RoleScope.COURSE, role_type=RoleType.LEARNER
+        )
+
+        stats = self.h_api.get_annotation_counts(
+            # Annotations in the course group and any children
+            [course.authority_provided_id]
+            + [child.authority_provided_id for child in course.children],
+            group_by="assignment",
+            h_userids=[s.h_userid for s in course_students],
+        )
+        # Organize the H stats by assignment ID for quick access
+        stats_by_assignment = {s["assignment_id"]: s for s in stats}
+        assignments: list[APIAssignment] = []
+
+        # Same course for all these assignments
+        api_course = APICourse(id=course.id, title=course.lms_name)
+        for assignment in self.course_service.get_assignments(
+            course, h_userid=self.request.user.h_userid if self.request.user else None
+        ):
+            if h_stats := stats_by_assignment.get(assignment.resource_link_id):
+                metrics = AnnotationMetrics(
+                    annotations=h_stats["annotations"],
+                    replies=h_stats["replies"],
+                    last_activity=h_stats["last_activity"],
+                )
+            else:
+                # Assignment with no annos, zeroing the stats
+                metrics = AnnotationMetrics(
+                    annotations=0, replies=0, last_activity=None
+                )
+
+            assignments.append(
+                APIAssignment(
+                    id=assignment.id,
+                    title=assignment.title,
+                    course=api_course,
+                    annotation_metrics=metrics,
+                )
+            )
+
+        return {"assignments": assignments}
